# Fix: Optimize network stack for <5ms ping performance