… .gradle/8.1.1/checksums/checksums.lock
… .gradle/8.1.1/checksums/md5-checksums.bin
… .gradle/8.1.1/checksums/sha1-checksums.bin
… .gradle/8.1.1/dependencies-accessors/dependencies-accessors.lock
… .gradle/8.1.1/executionHistory/executionHistory.bin
… .gradle/8.1.1/executionHistory/executionHistory.lock
… .gradle/8.1.1/fileHashes/fileHashes.bin
… .gradle/8.1.1/fileHashes/fileHashes.lock
… .gradle/buildOutputCleanup/buildOutputCleanup.lock
… docs/UART_PIN_CONFIGURATION.md
… include/network/lwip_netif_enc28j60.h
… include/network/network_manager.h
… src/core1_main.c
… src/main.c
… src/network/CMakeLists.txt
… src/network/enc28j60_driver.c
… src/network/lwip_netif_enc28j60.c
… src/network/network_manager.c
… tests/CMakeLists.txt
… tests/network/CMakeLists.txt
… tests/network/test_dhcp_debug.c
… tests/network/test_enc28j60_driver.c
… tests/network/test_network_integration.c